requirements/prod.txt
bleach==6.3.0
dj-database-url==3.1.0
Django==5.2.7
django-ckeditor-5==0.2.18
django-cleanup==9.0.0
django-environ==0.12.0
django-multiupload==0.6.1
django-tz-detect==0.5.0
pillow==12.0.0
psycopg2-binary==2.9.11
sorl-thumbnail==12.11.0
tinycss2==1.5.1
